✨ Feat: User able to decide to use current machine ip or localhost as the host address
Add localhost as a parameter in the constructor the user able to switch between using current
machine ip or localhost as the host address. If localhost = true, then it will use localhost as the
host name. If localhost = false, then it will use the current ip addres as the host address for
webpack-dev-server. This configuration allow user to use webpack-dev-server as  a debug server for
other machine.